client/requirements.txt
monitorcontrol==3.1.0
pyserial==3.5
